Run tests as a non-root user
Looks like the tests were failing because they were expecting to be run
as a non-privileged user, asserting that things behaved appropriately if
we tried to write a file but couldn't. Turns out, `root` can do whatever
it wants! :P

We may want to push this user creation down into a lower-level image;
not quite sure just yet.

Adds a `.dockerignore` file, as well as a `Makefile` with a targets that
allow for running tests both locally as well as within a Docker
container.